Refactor divider tests into own file
